BugFix partial correct of a+b+ behaviour
now updates equation, number,  & answer correctly, leaves main display clear. .... may be less confusing this way, actually. Evaluate if original intent is the better way (to display formula of a+b=ANS + AND the Ans in main display).
anything = does not update equation display or equation. does otherwirse behave correctly.

Multiple changes, biggest one of which is having  calcEquation() return an object - no attempts at *clearing* this object yet. I beleive it should not be recalled on next use of calcEquation() because defined in the scope of this function.