BugFix ANS,Num,+ & 'Impossible' assign correctly
fixed attempts to use impossible as a number.

For incorrect use of operator and number when pressed in number, operator order after calculating an answer with '=':
added call to updateEquation() inside calcEquation(), and removed updateEquation() after all instances of calcEquation.
In order to then tie clearing the 'operator' & 'numbers' variables to calcEquation() as well. This is the crux of fixing the problem where the operator is incorrectly applied to Ans + userEntry, when keyed in order as: ANS, userEntry, +