BugFix a+b== now returns correct number
a+b== would return the a+b initially, but the second = would show b=

The fix for this set off a chain of other minor bugs, but these have been taken care of.
Primarily by making sure userEntry is cleared every time '=' is called, and changing the if(numbers[0]) and if(numbers[1]) to numbers.length >= and === to force strict equality. Without this 0 values did not appear in the equation after fixes.