BugFix format a,+,BKSP,b,+ behaved correctly
Previously, when a number an operator had been entered, but the operator was backspaced, and a new number was entered followed by an operator, the calculator would use the initial number, the operator, and then new number as if they were entered in that order.

Desired behaviour is to reset the numbers variable if a new number is entered without an operator selected, to maintain consistency with the rest of the behaviour.