declare decimal & backspace btns as const
ids were added to the buttons to facilitate this.
The BKSP button now has id=bksp, and is referenced in const bkspBtn; the decimal id=nDot, and is referenced in const decimalBtn.

As an intermediate step towards toggling the states of these two buttons, as the previous plan would require deactivating the listener on all buttons and re-adding with/out the specific buttons. Seems inefficient & potentially complicated.

New plan is to avoid altering the listeners, and to route the functions differently based on class of the button. And of course, alter that via functions.