Create routing for deactivated '.' & 'BKSP'
altered the formulas for activating and deactivating
to set the class outright with its old value plus
'disabled' or not, instead of adding/removing, as
they may be called multiple times with no effect.
Removed initial disabled class from 'BKSP' in
index.html and added it back via JavaScript after
getting the class to facilitate this.

...realizing now that using an if(active) statement
would probably be more efficient. Oh well. This is
done and its not worth changing.